refactor displaying in html function as a module
